+++ prismaDmob/sql/getClientAllocatorDistributionWeekly.sql
@@ -3,7 +3,7 @@ select
     "addressId" as client,
     "verifierAddressId" as allocator,
     count(*)::int as num_of_allocations,
-    sum(allowance)::bigint as sum_of_allocations
+    sum(allowance)::float as sum_of_allocations
 from verified_client_allowance
 where height >= 3847920 -- nv22 start
 group by
